Убрал разделение на модули при открытии формы
